destroy 条件分岐追加

+++ app/controllers/items_controller.rb
@@ -37,10 +37,14 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    unless current_user == @item.user
+      redirect_to root_path
+    end
+
     if @item.destroy
       redirect_to root_path
     else
-      render :item
+      render :root
     end
   end
